requirements.txt
langchain_core
langchain_groq
flask
langchain
python-dotenv